requirements.txt
PyJWT
pandas
flask
gunicorn
python-dotenv
mysql-connector-python
pyotp
itsdangerous
qrcode[pil]